Migrate to isolated worker model
