+++ BookingQueueSubscriber/BookingQueueSubscriber.Common/BookingQueueSubscriber.Common.csproj
@@ -11,7 +11,6 @@
         <PackageReference Include="BookingsApi.Client" Version="2.0.5" />
         <PackageReference Include="LaunchDarkly.ServerSdk" Version="8.4.0" />
         <PackageReference Include="Microsoft.ApplicationInsights" Version="2.22.0" />
-        <PackageReference Include="Microsoft.Azure.Functions.Extensions" Version="1.1.0" />
         <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="8.0.0" />
         <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="8.0.0" />
         <PackageReference Include="Microsoft.IdentityModel.Clients.ActiveDirectory" Version="5.2.9" />
