+++ BookingQueueSubscriber/BookingQueueSubscriber/BookingQueueSubscriberFunction.cs
@@ -1,21 +1,12 @@
 namespace BookingQueueSubscriber
 {
-    public class BookingQueueSubscriberFunction
+    public class BookingQueueSubscriberFunction(IMessageHandlerFactory messageHandlerFactory, ILogger<BookingQueueSubscriberFunction> logger)
     {
-        private readonly IMessageHandlerFactory _messageHandlerFactory;
-        private readonly ILogger<BookingQueueSubscriberFunction> _logger;
-
-        public BookingQueueSubscriberFunction(IMessageHandlerFactory messageHandlerFactory, ILogger<BookingQueueSubscriberFunction> logger)
-        {
-            _messageHandlerFactory = messageHandlerFactory; 
-            _logger = logger;
-        }
-
-        [FunctionName("BookingQueueSubscriberFunction")]
+        [Function("BookingQueueSubscriberFunction")]
         public async Task Run([ServiceBusTrigger("%queueName%", Connection = "ServiceBusConnection")]
             string bookingQueueItem)
         {
-            _logger.LogInformation("Processing message {BookingQueueItem}", bookingQueueItem);
+            logger.LogInformation("Processing message {BookingQueueItem}", bookingQueueItem);
             // get handler
             EventMessage eventMessage;
             try
@@ -24,15 +15,15 @@ namespace BookingQueueSubscriber
             }
             catch (Exception e)
             {
-                _logger.LogCritical(e, "Unable to deserialize into EventMessage \r\n {BookingQueueItem}", bookingQueueItem);
+                logger.LogCritical(e, "Unable to deserialize into EventMessage \r\n {BookingQueueItem}", bookingQueueItem);
                 throw;
             }
 
-            var handler = _messageHandlerFactory.Get(eventMessage.IntegrationEvent);
-            _logger.LogInformation("using handler {Handler}", handler.GetType());
+            var handler = messageHandlerFactory.Get(eventMessage.IntegrationEvent);
+            logger.LogInformation("using handler {Handler}", handler.GetType());
 
             await handler.HandleAsync(eventMessage.IntegrationEvent);
-            _logger.LogInformation("Process message {EventMessageId} - {EventMessageIntegrationEvent}", eventMessage.Id,
+            logger.LogInformation("Process message {EventMessageId} - {EventMessageIntegrationEvent}", eventMessage.Id,
                 eventMessage.IntegrationEvent);
         }
     }
