+++ BookingQueueSubscriber/BookingQueueSubscriber/GlobalUsings.cs
@@ -21,9 +21,7 @@ global using BookingQueueSubscriber.Services.MessageHandlers.Core;
 global using Microsoft.ApplicationInsights.Extensibility;
 global using Microsoft.AspNetCore.Http;
 global using Microsoft.AspNetCore.Mvc;
-global using Microsoft.Azure.Functions.Extensions.DependencyInjection;
-global using Microsoft.Azure.WebJobs;
-global using Microsoft.Azure.WebJobs.Extensions.Http;
+global using Microsoft.Azure.Functions.Worker;
 global using Microsoft.Extensions.Caching.Memory;
 global using Microsoft.Extensions.Configuration;
 global using Microsoft.Extensions.DependencyInjection;
