+++ BookingQueueSubscriber/BookingQueueSubscriber/host.json
@@ -3,11 +3,9 @@
   "extensions": {
     "serviceBus": {
       "prefetchCount": 16,
-      "messageHandlerOptions": {
-        "autoComplete": true,
-        "maxConcurrentCalls": 32,
-        "maxAutoRenewDuration": "00:05:00"
-      }
+      "autoCompleteMessages": true,
+      "maxConcurrentCalls": 32,
+      "maxAutoLockRenewalDuration": "00:05:00"
     },
     "http": {
       "routePrefix": "",
